pyproject.toml
[tool.poetry]
name = "course-7"
version = "0.1.0"
description = ""
authors = ["Aleksey <oszn2011@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
celery = "^5.4.0"
redis = "^5.1.0"
flake8 = "^7.1.1"
pillow = "^10.4.0"
black = "^24.8.0"
django-phonenumber-field = "^8.0.0"
rest-framework-simplejwt = "^0.0.2"
djangorestframework-simplejwt = "^5.3.1"
django-celery-beat = "^2.7.0"
drf-yasg = "^1.21.7"
setuptools = "^75.1.0"
phonenumbers = "^8.13.47"
django-countries = "^7.6.1"
python-telegram-bot = "^21.6"
coverage = "^7.6.1"
django-cors-headers = "^4.4.0"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = "migrations"